after now takes float seconds as argument
The after statement has thus far taken an integer argument for the
number of seconds after which to run a method. It is now changed to a
float so we can do sub-second delays.

Remove old test case that showed how we do not automatically convert
integers to float. Since it's not really a bug but rather a design
limitation we don't need a regression test for it.

+++ rts/rts.c
@@ -815,7 +815,7 @@ $Msg $ASYNC($Actor to, $Cont cont) {
     return m;
 }
 
-$Msg $AFTER($int sec, $Cont cont) {
+$Msg $AFTER($float sec, $Cont cont) {
     $Actor self = ($Actor)pthread_getspecific(self_key);
     rtsd_printf("# AFTER by %ld", self->$globkey);
     time_t baseline = self->$msg->$baseline + sec->val * 1000000;
